+++ packages/core/src/llm/azure.ts
@@ -17,9 +17,14 @@ const ALL_AZURE_OPENAI_CHAT_MODELS = {
     contextWindow: 16384,
     openAIModel: "gpt-3.5-turbo-16k",
   },
+  "gpt-4o": { contextWindow: 128000, openAIModel: "gpt-4o" },
   "gpt-4": { contextWindow: 8192, openAIModel: "gpt-4" },
   "gpt-4-32k": { contextWindow: 32768, openAIModel: "gpt-4-32k" },
   "gpt-4-turbo": {
+    contextWindow: 128000,
+    openAIModel: "gpt-4-turbo",
+  },
+  "gpt-4-turbo-2024-04-09": {
     contextWindow: 128000,
     openAIModel: "gpt-4-turbo",
   },
@@ -30,6 +35,10 @@ const ALL_AZURE_OPENAI_CHAT_MODELS = {
   "gpt-4-1106-preview": {
     contextWindow: 128000,
     openAIModel: "gpt-4-1106-preview",
+  },
+  "gpt-4o-2024-05-13": {
+    contextWindow: 128000,
+    openAIModel: "gpt-4o-2024-05-13",
   },
 };
 
@@ -62,6 +71,8 @@ const ALL_AZURE_API_VERSIONS = [
   "2024-02-01",
   "2024-02-15-preview",
   "2024-03-01-preview",
+  "2024-04-01-preview",
+  "2024-05-01-preview",
 ];
 
 const DEFAULT_API_VERSION = "2023-05-15";