+++ packages/autotool/examples/02_nextjs/CHANGELOG.md
@@ -1,5 +1,24 @@
 # @llamaindex/autotool-02-next-example
 
+## 0.1.7
+
+### Patch Changes
+
+- Updated dependencies [6e156ed]
+- Updated dependencies [265976d]
+- Updated dependencies [8e26f75]
+  - llamaindex@0.3.15
+  - @llamaindex/autotool@0.0.1
+
+## 0.1.6
+
+### Patch Changes
+
+- Updated dependencies [6ff7576]
+- Updated dependencies [94543de]
+  - llamaindex@0.3.14
+  - @llamaindex/autotool@0.0.1
+
 ## 0.1.5
 
 ### Patch Changes